==> scripts/bases/prodes/prodes_Caatinga/code/03_crossCheck.R <==
# SETUP ----------------------------------------------------------------------------------------------------------------------------------------------
rm(list = ls())
gc()

str.time <- Sys.time()

# GLOBAL SETTINGS
source("config.R")  # sets local dirs and sources config for shared data repo
source(file.path("_functions", "associateCRS.R"))
source(file.path("_functions", "prevalent_values.R"))
source(file.path("_functions", "convertUnits.R"))


# loading packages
pkgs <- c("tidyverse","sf", "rlang","future.apply","furrr", "parallel", 
          "tictoc","labelled", "geobr", "dplyr","openxlsx")

# groundhogLibraries(pkgs, "2022-04-21"
#                    # , tolerate.R.version =  "4.2.2"
#                    )

groundhogLibraries(pkgs, "2022-04-21"
                   # , tolerate.R.version =  "4.2.2"
)

DIR.DATA.PROJECT <- "A:/projects/car_a_car"

# Turns off scientific notation
options(scipen = 99) 

defo.biome <- file.path(
  DIR.CPI.DATA,
  "land/prodes_caatinga/cleanData/02_mrg_caatingaMuni"
)


### Calculate yearly defo crossed with municipalities ####

defo.files <- list.files(defo.biome)
munis <- str_extract(defo.files,pattern = "\\d{7}")



calculate.defo <- function(municipality) {
  
  load(
    file.path(
      defo.biome,
      paste0("prodes_inc_caat_muni_",
             municipality,
             "_sf.Rdata")
    )
  )
  
  result <- result %>% 
    mutate(defo_sqkm = as.numeric(st_area(geometry))*convert.sqm.to.sqkm) %>% 
    as_tibble() %>% 
    select(-geometry) %>% 
    group_by(prodes_year) %>% 
    summarize(defo_sqkm = sum(defo_sqkm))
  
}


plan(multisession, workers = detectCores()-1)

res = future_lapply(munis,calculate.defo)

answer <- bind_rows(res) %>% 
  group_by(prodes_year) %>% 
  summarize(crossedMuni = sum(defo_sqkm))


### calculate defo before crossing #####

## Load Prodes Increment ####

aux.name <- load(
  file.path(
    DIR.CPI.DATA,
    "land",
    "prodes_caatinga\\cleanData",
    "cln_lcv_dfrst_caat_biome_prodes_inpe_sf.Rdata"
  )
)

#get the object by its names
prodes_increments <- get(aux.name)

#remove the old object
rm(list = aux.name, aux.name)


answer <-   prodes_increments %>% as_tibble() %>% select(-geometry) %>%
  mutate(prodes_year = as.double(prodes_year)) %>%
  group_by(prodes_year) %>%
  summarize(asWritten = sum(prodes_area_km2)) %>%
  right_join(answer, by = "prodes_year")

answer = answer %>% mutate(diff =  crossedMuni - asWritten)

sum(answer$diff)

#########################################################


write.xlsx(answer,file.path(
  DIR.CPI.DATA,
  "land/prodes_Caatinga/cleanData","04_crossCheck.xlsx"))

print(
  Sys.time() - str.time
)
